The great migration >:3
All branches were merged to master when this went down.
I thought it was as good a time as any to organize the repo.
If anyone disagrees with this structure, please flame the wiki.

I tried to keeps all code edits logical. It builds! :O
Nothing was modified that didn't have to be. Caveat:
To avoid duplicate compilations, I packaged message in a lib.